feat: export metrics for Prometheus with OpenTelemetry instrumentation
